remove workflow execution module
